+++ jeepay-service/src/main/java/com/jeequan/jeepay/service/impl/MchAppService.java
@@ -36,7 +36,7 @@ public class MchAppService extends ServiceImpl<MchAppMapper, MchApp> {
     public void removeByAppId(String appId) {
 
         // 1.查看当前应用是否存在交易数据
-        int payCount = payOrderService.count(PayOrder.gw().eq(PayOrder::getAppId, appId));
+        long payCount = payOrderService.count(PayOrder.gw().eq(PayOrder::getAppId, appId));
         if (payCount > 0) {
             throw new BizException("该应用已存在交易数据，不可删除");
         }
